+++ CHANGELOG.md
@@ -5,7 +5,8 @@
   - line and block comments
   - indentation-based folding ranges.
 - Add `sorbetto.rubyFmtPath` setting to configure `rubyfmt` path.
-- Fix: Missing `Initializing` status in some cases, making it unclear whether the language server is starting up.
+- Improve `sorbetto.restartFilePatterns` restart-listener now remain active after error, allowing on-disk changes to `sorbet/config` to take effect immediately.
+- Fix: Missing **Initializing** message in language status item in some scenarios, making it unclear whether the language server is starting up.
 
 ## 0.3.21
 - Fix: Disable `FoldingRangeProvider` for `#` comment blocks as it disables VS Code's indentation-based folding logic.